MF-000 Extract MetroLineLayers from MetroMap
